Work in progress on non-rectangular top surface integration during cell creation. Moved quad and vertex into their own files for import because my new functions in shapely_utils.py need to import quad but they can't import existing quad from grid_tesselate due to a circular reference since the new functions are called from grid_tesselate. Simplified the code for marking walls and handle case where cell1 or cell2 are disjoint or contained properly in boundary.
